@@ -24,8 +24,8 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
     }
     namespace 'com.stadtiq' // <<< CHANGE THIS >>>
 }
@@ -36,6 +36,7 @@ dependencies {
     implementation libs.material // Or the latest stable version
     implementation libs.constraintlayout // Or your preferred layout constraint version
     implementation libs.recyclerview // Or the latest stable version
+    implementation libs.mpandroidchart
 
     // Required for SharedPreferences handling in a clean way
     implementation libs.preference // Or the latest stable version
